REFACTOR: PR 코드리뷰 Warning 3건 수정
- formatRelativeTime 유틸 함수를 shared/utils/date.ts로 추출하여 FeedCard, PostDetail 중복 제거
- PostDetail에 DOMPurify sanitize 적용으로 XSS 취약점 방지
- api-types/server.ts에서 .extend() 패턴을 인라인 z.object로 통합
