+++ apps/client/package.json
@@ -34,6 +34,7 @@
     "@trpc/react-query": "11.4.3",
     "axios": "^1.7.9",
     "ckeditor5": "^47.4.0",
+    "dompurify": "^3.3.1",
     "lucide-react": "^0.575.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
