+++ apps/client/src/components/feed/FeedCard.tsx
@@ -2,6 +2,7 @@ import { Link } from '@tanstack/react-router';
 import tw from 'tailwind-styled-components';
 
 import FollowButton from '../follow/FollowButton';
+import { formatRelativeTime } from '../../shared/utils/date';
 
 interface FeedCardProps {
   postId: string;
@@ -13,23 +14,6 @@ interface FeedCardProps {
   thumbnail?: string | null;
   publishedAt: string | Date | null;
   tags?: string[];
-}
-
-function formatRelativeTime(date: string | Date | null): string {
-  if (!date) return '';
-  const now = new Date();
-  const target = new Date(date);
-  const diffMs = now.getTime() - target.getTime();
-  const diffMin = Math.floor(diffMs / 60000);
-  if (diffMin < 1) return '방금 전';
-  if (diffMin < 60) return `${diffMin}분 전`;
-  const diffHour = Math.floor(diffMin / 60);
-  if (diffHour < 24) return `${diffHour}시간 전`;
-  const diffDay = Math.floor(diffHour / 24);
-  if (diffDay < 30) return `${diffDay}일 전`;
-  const diffMonth = Math.floor(diffDay / 30);
-  if (diffMonth < 12) return `${diffMonth}개월 전`;
-  return `${Math.floor(diffMonth / 12)}년 전`;
 }
 
 const FeedCard = ({
